"Checktab" in m_impurityoperatoroffdiag modified for abiref_nag_6.2_openmpi

+++ src/63_ctqmc_offdiag/m_ImpurityOperatoroffdiag.F90
@@ -2025,7 +2025,7 @@ SUBROUTINE ImpurityOperatoroffdiag_checkOverlap(op, Tmin, Tmax, iOverlap, iflavo
   INTEGER                                :: imaxbeta
   INTEGER                                :: isegment
   INTEGER                                :: tail
-  BYTE, DIMENSION(1:size,1:2)            :: checktab
+  INTEGER(1), DIMENSION(1:size,1:2)      :: checktab 
   CHARACTER(LEN=4)                       :: a
   DOUBLE PRECISION                       :: dt
   DOUBLE PRECISION                       :: inv_dt
